T-squared Confidence Ellipse Coordinates

This function produces coordinates values for plotting confidence ellipse.

ellipseCoord(data, pcx = 1, pcy = 2, conf.limit = 0.95, pts = 200)

Arguments

data
a data frame or tibble of scores coordinates

pcx
an integer specifying which component is on the x-axis (by default 1)

pcy
an integer specifying which component is on the y-axis (by default 2)

conf.limit
an integer specifying the confidence level (by default 0.95)

pts
an integer indicating the number of points in the ellipse (by default 200)

Value

a data frame with x and y coordinates

Details

Coordinates for Hotelling's confidence ellipse

Author

Christian L. Goueguel, christian.goueguel@gmail.com

Examples

## Principal components analysis (PCA)
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

set.seed(123)
pca_mod <- specData %>%
  dplyr::select(where(is.numeric)) %>%
  FactoMineR::PCA(scale.unit = FALSE, graph = FALSE)

## Extract PCA scores
pca_scores <- pca_mod %>%
   purrr::pluck("ind", "coord") %>%
   tibble::as_tibble()

## Compute Hotelling T-squared statistic and ellipse parameters
library(HotellingEllipse)
xy_coord <- ellipseCoord(data = pca_scores, pcx = 1, pcy = 2, conf.limit = 0.95, pts = 200)